requirements.txt
paho-mqtt
aiocoap
smbus2
Adafruit_DHT
RPi.GPIO
matplotlib
numpy
scipy
